list: drop legacy runner support
